move log out button to profile screen and remove profile btn from home screen
