+++ src/screens/UserProfileScreen.js
@@ -2,6 +2,7 @@ import React from "react";
 import { View, Text, FlatList, TouchableOpacity } from "react-native";
 import { useUser } from "../context/UserContext";
 import firebaseApp from "../utils/firebase";
+import { getAuth } from "firebase/auth";
 import { getDatabase, ref, onValue } from "firebase/database";
 import useList from "../hooks/useList";
 import { globalStyles } from "../styles/global";
@@ -11,8 +12,17 @@ import ListDetail from "../components/ListDetailComponent";
 // list favorites
 
 export default function UserProfile({ navigation }) {
-  const { uid, username } = useUser();
+  const auth = getAuth(firebaseApp);
+  const { uid, username, setUsername, setIsLoggedIn, setFavs } = useUser();
   const [coffeeList] = useList();
+
+  const handleSignout = () => {
+    auth.signOut();
+    setIsLoggedIn(false);
+    setUsername("");
+    setFavs({});
+    navigation.navigate("Home");
+  };
 
   const db = getDatabase(firebaseApp);
   const userRef = ref(db, "users/" + uid);
@@ -75,6 +85,11 @@ export default function UserProfile({ navigation }) {
         <Text style={globalStyles.plainText}>Your favorite cafés: </Text>
       </View>
       {favlist}
+      <View style={globalStyles.buttonContainer}>
+        <TouchableOpacity onPress={handleSignout}>
+          <Text style={globalStyles.touchable}>Log out</Text>
+        </TouchableOpacity>
+      </View>
     </View>
   );
 }
